more maintainability improvements
