+++ .github/workflows/test.yml
@@ -12,4 +12,8 @@ jobs:
           node-version: 24
           cache: npm
       - run: npm ci
+      - run: npm run lint
       - run: npm test
+      - run: npm run build
+      - name: Check fixtures/ is up to date
+        run: git diff --exit-code fixtures/
